requirements.txt
fastapi
uvicorn
scikit-learn
joblib
python-dotenv
xgboost
pandas
